rename display property and create table content shared service
