+++ src/app/app.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { IamService } from 'src/library/open-services/modules/iam/iam.service';
 import { DEFAULT_SETTING } from '../shared/constant/displayPropertiesContant';
-import { DisplayPropertiesService } from 'src/shared/services/display-properties/display-properties.service';
+import { DisplayPropertiesService } from 'src/shared/services/table-header-display-properties/display-properties.service';
 
 @Component({
   selector: 'app-root',
