+++ src/app/modules/cognito/pages/cognito-apps/cognito-apps.component.ts
@@ -7,7 +7,7 @@ import { MatDialog, MatButtonToggleChange } from '@angular/material';
 import { ConfirmationModal, ManageColumnModal } from '@perx/open-ui-components';
 import { map } from 'rxjs/operators';
 import { TableHeaderProperties } from 'src/shared/models/tableHeaderProperties';
-import { DisplayPropertiesService } from 'src/shared/services/display-properties/display-properties.service';
+import { DisplayPropertiesService } from 'src/shared/services/table-header-display-properties/display-properties.service';
 
 
 @Component({
