+++ src/app/modules/comms/pages/campaigns/new-campaign/new-campaign.component.ts
@@ -5,7 +5,7 @@ import { CommsService, CognitoService, CognitoPool } from '@perx/open-services';
 import { map, takeUntil } from 'rxjs/operators';
 import { Observable, Subject } from 'rxjs';
 import { TableHeaderProperties } from 'src/shared/models/tableHeaderProperties';
-import { DisplayPropertiesService } from 'src/shared/services/display-properties/display-properties.service';
+import { DisplayPropertiesService } from 'src/shared/services/table-header-display-properties/display-properties.service';
 
 @Component({
   selector: 'app-new-campaign',
