+++ src/app/modules/comms/pages/events/new-event/new-event.component.ts
@@ -14,7 +14,7 @@ import { SelectionModel } from '@angular/cdk/collections';
 import { map, takeUntil } from 'rxjs/operators';
 import { NgxMaterialTimepickerComponent } from 'ngx-material-timepicker';
 import { TableHeaderProperties } from 'src/shared/models/tableHeaderProperties';
-import { DisplayPropertiesService } from 'src/shared/services/display-properties/display-properties.service';
+import { DisplayPropertiesService } from 'src/shared/services/table-header-display-properties/display-properties.service';
 
 
 @Component({
