+++ src/app/modules/iam/pages/iam-groups/iam-groups.component.html
@@ -20,22 +20,16 @@
       Detach Groups<span style="vertical-align: unset" *ngIf="selection && selection.selected.length>1">s</span> from User
     </button>
 
-    <div fxFlex fxLayout="row" fxLayoutAlign="end">
-      <mat-button-toggle-group
-          aria-label="Other Actions"
-          name="otherActions"
-          (change)="onOtherActionsChange($event)">
-        <mat-button-toggle value="reload">
-          <mat-icon>refresh</mat-icon>
-        </mat-button-toggle>
-        <mat-button-toggle value="settings">
-          <mat-icon>settings</mat-icon>
-        </mat-button-toggle>
-        <mat-button-toggle value="help">
-          <mat-icon>help</mat-icon>
-        </mat-button-toggle>
-      </mat-button-toggle-group>
-    </div>
+    <app-table-actions-management
+      fxFlex fxLayout="row" fxLayoutAlign="end"
+      [tablePlatform]="'essentials'"
+      [tableModule]="'IAM'"
+      [tableName]="'groups-table'"
+      [shownColumns]="shownColumns"
+      (reloadTableEmit)="reloadTable()"
+      (changeTableHeaderSettingEmit)="changeTableHeaderSetting($event)"
+    >
+    </app-table-actions-management>
   </div>
   <ng-container class="groups-table" *ngIf="groups$ | async as groups">
     <app-filterable-table
