+++ src/library/open-ui-components/components/table-actions-management/table-actions-management.component.html
@@ -1,17 +1,11 @@
-<div>
-  <mat-button-toggle-group
-      aria-label="Other Actions"
-      name="otherActions"
-      (change)="onOtherActionsChange($event)"
-  >
-    <mat-button-toggle value="reload">
-      <mat-icon>refresh</mat-icon>
-    </mat-button-toggle>
-    <mat-button-toggle value="settings">
-      <mat-icon>settings</mat-icon>
-    </mat-button-toggle>
-    <mat-button-toggle value="help">
-      <mat-icon>help</mat-icon>
-    </mat-button-toggle>
-  </mat-button-toggle-group>
-</div>
+<mat-button-toggle-group aria-label="Other Actions" name="otherActions" (change)="onOtherActionsChange($event)">
+  <mat-button-toggle value="reload">
+    <mat-icon>refresh</mat-icon>
+  </mat-button-toggle>
+  <mat-button-toggle value="settings">
+    <mat-icon>settings</mat-icon>
+  </mat-button-toggle>
+  <mat-button-toggle value="help">
+    <mat-icon>help</mat-icon>
+  </mat-button-toggle>
+</mat-button-toggle-group>
